Moved physics scripts higher up the hierarchy. map-builder is now among the map scripts.
